Raise `UnrecognizedAttributes` before other errors
When passing attributes with string keys, the first error you get looks
something like:

    ValueSemantics::MissingAttributes: Attribute `X#y` has no value

At which point you say "wtf, I did give a value for that". This happens
because the string keys are not detected as being associated with the
attributes.

Now that `ValueSemantics::UnrecognizedAttributes` gets raised, it should
be easier to diagnose the problem as being string keys.